rename io2 as iofunc
